requirements.txt
rdkit-pypi>=2022.9.5
pandas>=1.5.0
numpy>=1.23.0,<2.0.0
scikit-learn>=1.2.0
xgboost>=1.7.0
shap>=0.41.0
streamlit>=1.20.0
matplotlib>=3.6.0
seaborn>=0.12.0
joblib>=1.2.0
tqdm>=4.65.0
imbalanced-learn>=0.10.0
plotly>=5.13.0
ipykernel>=6.20.0
jupyter>=1.0.0
optuna>=3.0.0
torch>=2.0.0
torch-geometric>=2.3.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
